requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
python-jose
passlib[bcrypt]
python-dotenv
email-validator